Move hoverWidget to hoverService
