+++ src/vs/editor/browser/services/hoverService/hover.css
