+++ src/vs/editor/browser/services/hoverService/hoverWidget.ts
